Fix cache issue
The code was able to modify the original post Object before it was saved
to the cache. I've patched by cloning the post into a new object that
isn't later cached.

A side-effect was that the next post search couldn't find the cloned
object in the original array, so I've also made a getPostIndex()
function to do the work of indexOf...